feat: Complete VectorBT migration for all feature categories
